Fix 204 No Content response on registration DELETE
Since a 204 cannot have a message body [1] this sets the DELETE on registrations
to use `w.WriteHeader(204)` directly, vs the JSON helper.

Requests do work and successfully return a 204 on this currently, however an
error is logged while writing the response in the application:

```
ERROR handlers/helpers.go:30 error writing response {"error": "http: request method or response status code does not allow body"}
```

This also adds some explicit expectations to the response body in the tests.

[1] https://www.rfc-editor.org/rfc/rfc7231#section-6.3.5