Update internal user mock for OCM/AMS
Having a dynamic `org_id` means that it can cause mismatches in looking up
registrations when mocks are being used, since that `org_id` value is not static.

This just sets it to the username, since we only really care that it's a string value.
This will ensure at least per username, the org will be the same.