Fix `is_internal` assignment
We recently were able to configure test accounts with the labels we expect in a
deployed environment, and that requires tweaking our expectations on the label
object checks a bit.

The `Key()` of a label is the actual label name we reference, where the `Value()`
we've set as the truthiness of this internal label. Though we'll likely only ever
apply this label with a `"true"` value, we should still add an explicit check on
the key and the value, vs just the existence of the environmental key.