Adds support for Keycloak JWT module
Since Keycloak and Cognito follow the JWK standard [1], we can reuse the same
logic as defined in the `awsModule`. For the sake of being explicit, this adds
another module `keycloakModule`, so that in the future if any module-specific
needs arise, we can easily break them out, and it will also make it clear which
JWK/JWT verification backing is used in an environment.

[1] https://www.ietf.org/rfc/rfc7517.txt