Add default ordering direction and order by
Currently the UI implicitly assumes that the only field which can be ordered in
bop/mbop is `username`, so it's not explicitly passed in an `orderBy` param.

This will default to use `username` as the default order field for keycloak, and
also defaults the direction to `asc`, as the keycloak user service defaults to `desc`.

Once the UI supports multiple ordering fields, we'll need to adjust the supported
`orderBy` fields in the v3 endpoints and translate them for keycloak.